hotfix: Email content broken when restore app from background then swipe to next/previous email on tablet
