+++ lib/features/mailbox_dashboard/presentation/mailbox_dashboard_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:focus_detector_v2/focus_detector_v2.dart';
 import 'package:get/get.dart';
 import 'package:tmail_ui_user/features/email/presentation/email_view.dart';
+import 'package:tmail_ui_user/features/email/presentation/widgets/email_view_empty_widget.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/base_mailbox_dashboard_view.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/dashboard_routes.dart';
 import 'package:tmail_ui_user/features/search/email/presentation/search_email_view.dart';
@@ -38,7 +39,7 @@ class MailboxDashBoardView extends BaseMailboxDashBoardView {
             case DashboardRoutes.thread:
               return buildResponsiveWithDrawer(
                 left: ThreadView(),
-                right: const EmailView(),
+                right: const EmailViewEmptyWidget(),
                 mobile: buildScaffoldHaveDrawer(body: ThreadView()),
               );
