+++ lib/features/thread_detail/presentation/extension/handle_mark_multiple_emails_read_success.dart
@@ -3,7 +3,7 @@ import 'package:dartz/dartz.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email.dart';
 import 'package:model/email/read_actions.dart';
 import 'package:tmail_ui_user/features/thread_detail/presentation/extension/close_thread_detail_action.dart';
-import 'package:tmail_ui_user/features/thread_detail/presentation/extension/toggle_thread_detail_collape_expand.dart';
+import 'package:tmail_ui_user/features/thread_detail/presentation/extension/toggle_thread_detail_collapse_expand.dart';
 import 'package:tmail_ui_user/features/thread_detail/presentation/thread_detail_controller.dart';
 import 'package:tmail_ui_user/main/routes/route_navigation.dart';
 
@@ -26,7 +26,7 @@ extension HandleMarkMultipleEmailsReadSuccess on ThreadDetailController {
 
     final expandedEmail = emailIdsPresentation[expandedEmailId];
     if (emailIds.contains(expandedEmailId) && expandedEmail != null) {
-      toggleThreadDetailCollapeExpand(expandedEmail);
+      toggleThreadDetailCollapseExpand(expandedEmail);
     }
   }
 }
