+++ lib/features/thread_detail/presentation/thread_detail_manager.dart
@@ -82,7 +82,10 @@ class ThreadDetailManager extends ReloadableController {
           return;
         }
 
-        initializeThreadDetailManager(currentDisplayedEmails);
+        initializeThreadDetailManager(
+          currentDisplayedEmails,
+          selectedEmail,
+        );
       },
     );
   }
